Use "prepublish" instead of "prepublishOnly"
"prepublishOnly" is ignored by NPM <v4. Unfortunately, the
NPM in Ubuntu sources is 3.5.2, so it's quite likely
that "prepublishOnly" would be ignored when people run
"npm publish", which defeats its whole point.

"prepublish" is less than ideal because it's run *after* "npm install"
as well as *before* "npm publish" (god, I know). But in this case
that's probably acceptable.